feat : WaitingLineRepository 추상화
