+++ src/main/java/com/prgrms/catchtable/waiting/repository/waitingline/RedisWaitingLineRepository.java
@@ -18,7 +18,7 @@ import org.springframework.stereotype.Component;
 @Slf4j
 @RequiredArgsConstructor
 @Component
-public class RedisWaitingLineRepository {
+public class RedisWaitingLineRepository implements WaitingLineRepository{
 
     private final StringRedisTemplate redisTemplate;
